requirements.txt
cassandra-driver==3.11.0
Flask==0.12.2
Flask-RESTful==0.3.6
google-auth==1.0.2
google-cloud-core==0.26.0
google-cloud-storage==1.3.2
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
